requirements-cdk.txt
aws-cdk-lib==2.15.0
constructs>=10.0.0
boto3==1.24.80
aws-lambda-powertools==1.29.2